feat: rename package `model.auth` to `model.authentication` for consistency
Updated references project-wide as well.
