--- app/src/main/java/com/android/periodpals/model/auth/AuthModel.kt
@@ -1,22 +1,0 @@
-package com.android.periodpals.model.auth
-
-interface AuthModel {
-
-  suspend fun login(
-      userEmail: String,
-      userPassword: String,
-      onSuccess: () -> Unit,
-      onFailure: (Exception) -> Unit,
-  )
-
-  suspend fun register(
-      userEmail: String,
-      userPassword: String,
-      onSuccess: () -> Unit,
-      onFailure: (Exception) -> Unit,
-  )
-
-  suspend fun logout(onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
-
-  suspend fun isUserLoggedIn(token: String, onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
-}
