+++ app/src/main/java/com/android/periodpals/model/authentication/AuthModel.kt
@@ -1,0 +1,22 @@
+package com.android.periodpals.model.authentication
+
+interface AuthModel {
+
+  suspend fun login(
+    userEmail: String,
+    userPassword: String,
+    onSuccess: () -> Unit,
+    onFailure: (Exception) -> Unit,
+  )
+
+  suspend fun register(
+    userEmail: String,
+    userPassword: String,
+    onSuccess: () -> Unit,
+    onFailure: (Exception) -> Unit,
+  )
+
+  suspend fun logout(onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
+
+  suspend fun isUserLoggedIn(token: String, onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
+}
